workflow-examples/transaction-monitoring/backend-python/requirements.txt
flask==2.0.1
flask-cors==3.0.10
flask-sock==0.5.2
python-dotenv==0.19.0
requests==2.26.0
werkzeug==2.0.1 